requirements.txt
fastapi
uvicorn[standard]
sentence-transformers
faiss-cpu
requests
beautifulsoup4
numpy
pydantic
ruff
cloudscraper
umap-learn